chore: rename enter-repeat-agile-template for readability
